requirements.txt
fastapi
uvicorn
python-dotenv
streamlit
requests
pandas
openpyxl
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
chromadb